Prepare to remove ITransportFactory
