+++ src/se/vidstige/jadb/JadbConnection.java
@@ -3,31 +3,32 @@ package se.vidstige.jadb;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
+import java.net.Socket;
 import java.util.ArrayList;
 import java.util.List;
 
-public class JadbConnection {
+public class JadbConnection implements ITransportFactory {
 
-    private final ITransportFactory transportFactory;
+    private final String host;
+    private final int port;
+
+    private static final int DEFAULTPORT = 5037;
 
     public JadbConnection() {
-        this(new SocketTransportFactory());
-    }
-
-    public JadbConnection(ITransportFactory transportFactory) {
-        this.transportFactory = transportFactory;
+        this("localhost", DEFAULTPORT);
     }
 
     public JadbConnection(String host, int port) {
-        transportFactory = new SocketTransportFactory(host,port);
+        this.host = host;
+        this.port = port;
     }
 
-    final ITransportFactory getTransportFactory() {
-        return transportFactory;
+    public Transport createTransport() throws IOException {
+        return new Transport(new Socket(host, port));
     }
 
     public String getHostVersion() throws IOException, JadbException {
-        Transport main = transportFactory.createTransport();
+        Transport main = createTransport();
         main.send("host:version");
         main.verifyResponse();
         String version = main.readString();
@@ -37,7 +38,7 @@ public class JadbConnection {
 
     public InetSocketAddress connectToTcpDevice(InetSocketAddress inetSocketAddress)
             throws IOException, JadbException, ConnectionToRemoteDeviceException {
-        Transport transport = transportFactory.createTransport();
+        Transport transport = createTransport();
         try {
             return new HostConnectToRemoteTcpDevice(transport).connect(inetSocketAddress);
         } finally {
@@ -47,7 +48,7 @@ public class JadbConnection {
 
     public InetSocketAddress disconnectFromTcpDevice(InetSocketAddress tcpAddressEntity)
             throws IOException, JadbException, ConnectionToRemoteDeviceException {
-        Transport transport = transportFactory.createTransport();
+        Transport transport = createTransport();
         try {
             return new HostDisconnectFromRemoteTcpDevice(transport).disconnect(tcpAddressEntity);
         } finally {
@@ -56,7 +57,7 @@ public class JadbConnection {
     }
 
     public List<JadbDevice> getDevices() throws IOException, JadbException {
-        Transport devices = transportFactory.createTransport();
+        Transport devices = createTransport();
         devices.send("host:devices");
         devices.verifyResponse();
         String body = devices.readString();
@@ -65,7 +66,7 @@ public class JadbConnection {
     }
 
     public DeviceWatcher createDeviceWatcher(DeviceDetectionListener listener) throws IOException, JadbException {
-        Transport transport = transportFactory.createTransport();
+        Transport transport = createTransport();
         transport.send("host:track-devices");
         transport.verifyResponse();
         return new DeviceWatcher(transport, listener, this);
@@ -77,7 +78,7 @@ public class JadbConnection {
         for (String line : lines) {
             String[] parts = line.split("\t");
             if (parts.length > 1) {
-                devices.add(new JadbDevice(parts[0], parts[1], transportFactory));
+                devices.add(new JadbDevice(parts[0], parts[1], this));
             }
         }
         return devices;
